+++ instructions.h
@@ -2,22 +2,23 @@
 
 // #include <vector>
 #include<list>
+#include <cstdint>
 using namespace std;
 class instructions {
 public:
     // instruction's variables
-    unsigned long hex_add;
+    unsigned long long hex_add;
     short int type;
-    list<unsigned long> dependencies; // using vector because vector doesn't need to define its size when defining a vector
+    list<unsigned long long> dependencies; // using vector because vector doesn't need to define its size when defining a vector
     short int stage;
     bool getEX;
 
     
     instructions() { }
-    instructions(unsigned long add, short int t, list<unsigned long> deped) {
+    instructions(unsigned long long add, short int t, list<unsigned long long> deped) {
         hex_add = add;
         type = t;
-        for (unsigned long dependency : deped) {
+        for (unsigned long long dependency : deped) {
             dependencies.push_back(dependency);
         }
         stage = 0;
